+++ jenkins-universal-wrapper-pipeline.groovy
@@ -1894,11 +1894,13 @@ List actionSendReport(String actionLink, Map actionLinkItem, Object envVariables
                     mandatoryKeys[0], actionLink))
     mandatoryKeys += reportTarget == 'email' ? ['to'] : []
     mandatoryKeys += reportTarget == 'mattermost' ? ['url', 'text'] : []
-    List stringKeys = reportTarget == 'email' ? ['reply_to', 'subject', 'body'] : []
-    // TODO: pipeline setting change?
+    mandatoryKeys += reportTarget == 'telegram' ? ['bot_token', 'chat_id', 'text'] : []
+    List stringKeys = reportTarget == 'email' ? ['reply_to', 'subject', 'body'] : reportTarget == 'telegram' ?
+            ['message_thread_id', 'parse_mode', 'link_preview_options', 'api_url'] : []
+    List booleanKeys = reportTarget == 'telegram' ? ['disable_notification', 'protect_content'] : []
     (mandatoryKeyValues, actionLinkItem, newActionOk) = checkMandatoryKeysTemplateAndFilterMapWrapper(actionLinkItem,
-            mandatoryKeys, mandatoryKeys + stringKeys as ArrayList, [], newActionOk, check, actionLink, envVariables,
-            universalPipelineWrapperBuiltIns)
+            mandatoryKeys, mandatoryKeys + stringKeys as ArrayList, booleanKeys, newActionOk, check, actionLink,
+            envVariables, universalPipelineWrapperBuiltIns)
     String actionName = String.format('send report to %s', reportTarget.trim() ? reportTarget : '<undefined>')
     Closure actionClosure = mandatoryKeyValues[0] == 'email' ? {
         emailext(
@@ -1910,6 +1912,16 @@ List actionSendReport(String actionLink, Map actionLinkItem, Object envVariables
         [newActionOk, universalPipelineWrapperBuiltIns, null]
     } : mandatoryKeyValues[0] == 'mattermost' ? {
         Boolean sendReportStatus = CF.sendMattermostChannelSingleMessage(mandatoryKeyValues[1], mandatoryKeyValues[2],
+                getBooleanVarStateFromEnv(envVariables) ? 2 : 0)
+        [sendReportStatus && newActionOk, universalPipelineWrapperBuiltIns, null]
+    } : mandatoryKeyValues[0] == 'telegram' ? {
+        Map telegramData = actionLinkItem.findAll {
+            it.key != mandatoryKeys[0] && it.key != mandatoryKeys[1] && it.key != stringKeys[3]
+        }
+        if (telegramData?.get(stringKeys[0]))
+            telegramData[stringKeys[0]] = telegramData?.get(stringKeys[0])?.toInteger()
+        String telegramBotApiUrl = actionLinkItem?.get(stringKeys[3] as String) ?: GV.TELEGRAM_BOT_API_URL
+        Boolean sendReportStatus = CF.sendTelegramMessageViaBot(telegramData, mandatoryKeyValues[0], telegramBotApiUrl,
                 getBooleanVarStateFromEnv(envVariables) ? 2 : 0)
         [sendReportStatus && newActionOk, universalPipelineWrapperBuiltIns, null]
     } : {